pyproject.toml
[tool.poetry]
name = "passwordmanagerbe"
version = "0.1.0"
description = ""
authors = ["seanborg <sean_b@optimove.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
pyjwt = "^2.8.0"
cryptography = "^43.0.0"
uuid = "^1.30"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
